requirements.txt
pdftotext==2.1.2
refextract==0.2.16
python-Levenshtein==0.12.0
numpy==1.16.4